Pull dev to main

+++ src/app/api/documents/delete/[id]/route.ts
@@ -1,15 +1,17 @@
 import { NextResponse } from 'next/server';
-
 import { ObjectId } from 'mongodb';
 import { getCollection } from '@/app/utils/mongodb';
 
 const DATABASE_NAME = 'documents';
 const COLLECTION_NAME = 'fs.files';
 
-export async function DELETE(req: Request, { params }: { params: { id: string } }) {
-
+export async function DELETE({ params }: { params: { id: string } } ) {
     try {
         const { id } = await params;
+
+        if (!id) {
+            return NextResponse.json({ message: 'File ID is required' }, { status: 400 });
+        }
 
         const bucket = await getCollection(DATABASE_NAME, COLLECTION_NAME);
         const deleteFile = await bucket.deleteOne({ _id: new ObjectId(id) });
@@ -21,6 +23,6 @@ export async function DELETE(req: Request, { params }: { params: { id: string } 
         return NextResponse.json({ message: 'File deleted successfully' });
     } catch (error) {
         console.error('Error deleting file:', error);
-        return NextResponse.json({ message: 'Failed to delete file', error }, { status: 500 });
+        return NextResponse.json({ message: 'Failed to delete file', error: error instanceof Error ? error.message : error }, { status: 500 });
     }
 }
